+++ Sources/AWSLambdaRuntimeCore/Channel Handler/RuntimeAPIHandler.swift
@@ -22,7 +22,7 @@ enum RuntimeAPIRequest {
 }
 
 enum RuntimeAPIResponse {
-    case next(Lambda.Invocation, ByteBuffer)
+    case next(Invocation, ByteBuffer)
     case accepted
     case error(ErrorResponse)
 }
@@ -67,7 +67,7 @@ final class RuntimeAPIHandler: ChannelDuplexHandler {
                 return context.fireErrorCaught(Lambda.RuntimeError.invocationMissingHeader(AmazonHeaders.traceID))
             }
             
-            let invocation = Lambda.Invocation(
+            let invocation = Invocation(
                 requestID: requestID,
                 deadlineInMillisSinceEpoch: unixTimeInMilliseconds,
                 invokedFunctionARN: invokedFunctionARN,
@@ -162,4 +162,47 @@ internal enum AmazonHeaders {
     static let cognitoIdentity = "Lambda-Runtime-Cognito-Identity"
     static let deadline = "Lambda-Runtime-Deadline-Ms"
     static let invokedFunctionARN = "Lambda-Runtime-Invoked-Function-Arn"
+}
+
+internal struct ErrorResponse: Codable {
+    var errorType: String
+    var errorMessage: String
+}
+
+extension ErrorResponse {
+    internal func toJSONBytes() -> [UInt8] {
+        var bytes = [UInt8]()
+        bytes.append(UInt8(ascii: "{"))
+        bytes.append(contentsOf: #""errorType":"#.utf8)
+        self.errorType.encodeAsJSONString(into: &bytes)
+        bytes.append(contentsOf: #","errorMessage":"#.utf8)
+        self.errorMessage.encodeAsJSONString(into: &bytes)
+        bytes.append(UInt8(ascii: "}"))
+        return bytes
+    }
+}
+
+internal struct Invocation {
+    let requestID: String
+    let deadlineInMillisSinceEpoch: Int64
+    let invokedFunctionARN: String
+    let traceID: String
+    let clientContext: String?
+    let cognitoIdentity: String?
+    
+    init(
+        requestID: String,
+        deadlineInMillisSinceEpoch: Int64,
+        invokedFunctionARN: String,
+        traceID: String,
+        clientContext: String?,
+        cognitoIdentity: String?)
+    {
+        self.requestID = requestID
+        self.deadlineInMillisSinceEpoch = deadlineInMillisSinceEpoch
+        self.invokedFunctionARN = invokedFunctionARN
+        self.traceID = traceID
+        self.clientContext = clientContext
+        self.cognitoIdentity = cognitoIdentity
+    }
 }
